fixed the nav bar, able to get navigation working but needs to be permanent on all pages
