+++ dronuts/front-end/src/components/NavbarScroller.tsx
@@ -1,20 +1,38 @@
-import * as React from 'react';
-import { Link } from 'react-router-dom';
+import React from 'react';
+import { Tabs } from '@geist-ui/react';
+import { useLocation } from 'react-router-dom';
+import { useNavigate } from "react-router-dom";
 
-const NavbarScroller = (props: {
-  brand: { name: string; to: string },
-  links: Array<{ name: string, to: string }>
-}) => {
-  const { brand, links } = props;
-  const NavLinks: any = () => links.map((link: { name: string, to: string }) =>
-    <li key={link.name}><Link to={link.to}>{ link.name }</Link></li>
-  );
-  return (
-    <div>
-      <Link to="/">{ brand.name }</Link>
-      <NavLinks />
-    </div>
-  )
+
+//navigation ordering for the navbar
+const navigation = {
+  brand: { name: 'Home', to: '/' },
+  links: [
+    { name: 'Store', to: '/store' },
+    { name: 'Admin Store', to: '/adminstore' },
+    { name: 'Drones', to: '/drones' },
+    { name: 'Loading Instructions', to: '/loadinstructions' },
+    { name: 'Order Status', to: '/orderstatus' },
+    { name: 'Store Order History', to: '/storeorderhistory' },
+    { name: 'Employee Notification System', to: '/empnotsys' }
+  ]
 };
 
-export default NavbarScroller;
+const NavBarScroller = () => {
+  const location = useLocation();
+  const navigate = useNavigate();
+
+  return (
+    <Tabs value={location.pathname} onChange={(route) => navigate(route, { replace: true })}>
+      <Tabs.Item label="Store" value = "/store" /> 
+      <Tabs.Item label="Admin Store" value = "/adminstore" />
+      <Tabs.Item label="Loading Instructions" value = "/loadinstructions" />
+      <Tabs.Item label="Order Status" value = "/orderstatus" />
+      <Tabs.Item label="Store Order History" value = "/storeorderhistory" />
+      <Tabs.Item label="Employee Notification System" value = "/empnotsys" />
+      <Tabs.Item label="Drones" value = "/drones" />
+    </Tabs>
+  );
+};
+
+export default NavBarScroller;
